requirements.txt
flask
flask_jwt_extended
requests
python-dotenv
